+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 ~~~~~
 
 -  Custom user specific databases, use commands ``/save`` and ``/save_mode`` more information in ``/commands``
+-  Be able to show custom DB entries with ``/db_list``
 -  Add functionality to add alias commands just like a normal command but a string as ``command`` value, which points to
    a ``command_name``. Additionally ``title``, ``description``, ``hidden`` and ``group`` can be set.
 
@@ -34,6 +35,7 @@ Changes
 -  Reply to user message on GIF download, so that the user sees to which GIF the message belongs
 -  Improve TTS error message
 -  Rename ``tty`` command to ``tts`` (Text-To-Speech) but add an ``tty`` alias for the time being
+-  Be able to set a CallbackQueryHandler for a CallbackQuery sender
 
 
 [1.4.0] - 2018.05.18